[IMP] web_field_required_invisible_manager: compatibility with tier validation
