+++ web_field_required_invisible_manager/models/__init__.py
@@ -1,2 +1,4 @@
 from . import custom_field_restriction
-from . import models
+from . import base
+from . import ir_ui_view
+from . import ir_model
